Extract bitstream objects into their on header files
